Listar cuotas de pago de un estudiante y su estado(HU3)

+++ Preu_TopEducation_Ti/src/main/java/com/example/Preu_TopEducation_Ti/entities/CuotaEntity.java
@@ -16,13 +16,14 @@ public class CuotaEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idCuota;
-    private double cantidad;
+    private int cantidad;
     private double arancel;
     private double arancelMensual;
     private String fechaEmision;
     //private String fechaPago;
     private String estado;
 
+    // relacion mucho a uno del estudiante cuota //
     @ManyToOne
     @JoinColumn(name="rut")
     private EstudianteEntity estudiante;
